+++ lib/DueDatePicker/DueDateHeader.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import moment from 'moment';
 import cx from 'classnames';
-import DueDateTime from './time';
+import DueDateTime from './DueDateTime';
 
 const DueDateHeader = props => {
   if (props.dueDate) {
