+++ lib/DueDatePicker/index.js
@@ -6,8 +6,8 @@ import Popover from 'react-bootstrap/lib/Popover';
 import moment from 'moment';
 import cx from 'classnames';
 import Button from '../Button';
-import DueDateHeader from './header';
-import DueDateButton from './button';
+import DueDateHeader from './DueDateHeader';
+import DueDateButton from './DueDateButton';
 
 class DueDatePicker extends Component {
   constructor(props) {
